+++ ListableUI/Sources/Layout/ListLayout/LayoutDescription.swift
@@ -39,7 +39,7 @@ import Foundation
 /// Under the hood, Listable is smart, and will only re-create the underlying
 /// layout object when needed (when the layout type or layout appearance changes).
 ///
-public struct LayoutDescription
+public struct LayoutDescription : Equatable
 {
     let configuration : AnyLayoutDescriptionConfiguration
     
@@ -64,6 +64,10 @@ public struct LayoutDescription
     public var layoutAppearanceProperties : ListLayoutAppearanceProperties {
         configuration.layoutAppearanceProperties()
     }
+    
+    public static func == (lhs : Self, rhs : Self) -> Bool {
+        lhs.configuration.isEqual(to: rhs.configuration)
+    }
 }
 
 
@@ -84,11 +88,15 @@ extension ListLayout
 
 extension LayoutDescription
 {
-    public struct Configuration<LayoutType:ListLayout> : AnyLayoutDescriptionConfiguration
+    public struct Configuration<LayoutType:ListLayout> : AnyLayoutDescriptionConfiguration, Equatable
     {
         public let layoutType : LayoutType.Type
         
         public let layoutAppearance : LayoutType.LayoutAppearance
+        
+        public static func == (lhs: Self, rhs: Self) -> Bool {
+            lhs.layoutType == rhs.layoutType && lhs.layoutAppearance == rhs.layoutAppearance
+        }
         
         // MARK: AnyLayoutDescriptionConfiguration
         
@@ -141,6 +149,10 @@ extension LayoutDescription
             
             return self.layoutType == other.layoutType
         }
+        
+        public func isEqual(to other : AnyLayoutDescriptionConfiguration) -> Bool {
+            self == (other as? Self)
+        }
     }
 }
 
@@ -163,4 +175,6 @@ public protocol AnyLayoutDescriptionConfiguration
     func shouldRebuild(layout anyLayout : AnyListLayout) -> Bool
 
     func isSameLayoutType(as other : AnyLayoutDescriptionConfiguration) -> Bool
+    
+    func isEqual(to other : AnyLayoutDescriptionConfiguration) -> Bool
 }